Many Changes, XML cache, music, observers
Oberserver no longer gives error when trying to remove a callback when no callback is associated. This allows you to actually use the function UsubscribeKey(id).
XML now properly uses its cache to actually open up new files. but when you save a file, it will automatically tell the resource manager to remove that file from the cache.
The music was not properly looping because it was using properties = 0. not properties = GetTrackProperties. As properties is a type of map.
Added the fifth row to the Encounter info in order to not get a crash when trying to walk to the southern tip of the island